style: 💄 Apply PHP-CS-Fixer

+++ src/Concerns/SessionLocks.php
@@ -9,9 +9,9 @@ use Mpyw\LaravelDatabaseAdvisoryLock\Contracts\SessionLock;
 
 trait SessionLocks
 {
-    abstract public function lockOrFail(string $key, int|float $timeout = 0): SessionLock;
+    abstract public function lockOrFail(string $key, float|int $timeout = 0): SessionLock;
 
-    public function tryLock(string $key, int|float $timeout = 0): ?SessionLock
+    public function tryLock(string $key, float|int $timeout = 0): ?SessionLock
     {
         try {
             return $this->lockOrFail($key, $timeout);
